Separate PoS and PoW difficulty adjust
Set both PoW and PoS to adjust difficulty to nTargetTimespan each. This
gives an average block time of half nTargetTimespan. So nTargetTimespan
is 2 minutes on main chain, with PoS and PoW adjusting to produce one of
each of these every 2 minutes we'll see an average of a block a minute.